requirements.txt
Django
djangorestframework
prettyconf
requests
shapely
black
gunicorn
whitenoise
asyncio